Collapse useAcuant into document-capture test setup
