+++ spec/javascripts/packages/document-capture/components/document-capture-spec.jsx
@@ -15,9 +15,8 @@ import {
 import DocumentCapture, {
   except,
 } from '@18f/identity-document-capture/components/document-capture';
-import { render } from '../../../support/document-capture';
+import { render, useAcuant } from '../../../support/document-capture';
 import { useSandbox } from '../../../support/sinon';
-import { useAcuant } from '../../../support/acuant';
 
 describe('document-capture/components/document-capture', () => {
   const sandbox = useSandbox();
